// gen/api.ts
export interface Authorization {
    /**
     * 
     * @type {AuthorizationClient}
     * @memberof Authorization
     */
    client: AuthorizationClient;
    /**
     * 
     * @type {Array<string>}
     * @memberof Authorization
     */
    scopes: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof Authorization
     */
    audiences: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Authorization
     */
    grantedAt: string;
}
/**
 * 
 * @export
 * @interface AuthorizationClient
 */
export interface AuthorizationClient {
    /**
     * 
     * @type {string}
     * @memberof AuthorizationClient
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizationClient
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizationClient
     */
    tos?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizationClient
     */
    image?: string;
}